requirements.txt
pandas
numpy
scikit-learn
scipy
python-dotenv
